A notebook for question and answer generation using one of the most powerful opensource NLU models, FLAN-T5-11B. (#215)
* Create T5_closed_book_QA_generators.py

This is code that can be run in a notebook or by itself to generate a dictionary for use in creating synthetic dialogue that can be verified for factual accuracy. To use this notebook your need your trusted source material to be in the format of a list of strings (they will be truncated to under 1100 characters). Requires transformers and accelerate. Make sure to use T5 with bfloat16 or full precision.

* Create T5_closed_book_QA_generators.py

* Create T5_closed_book_QA_generators.md

* Delete T5_closed_book_QA_generators.py

Moved to proper folder structure

* fixed pre-commit issues

Co-authored-by: Yannic Kilcher <yk@users.noreply.github.com>