Introduce users and Discord auth to inference server (#1772)
Resolves #1471.

The idea here is to introduce a flexible system which allows arbitrarily
many authentication methods to be added in future, starting with
Discord. We use a Discord OAuth2 flow to login a user (or register a new
user), then issue a JSON Web Token (JWT) which can be decoded to obtain
a user ID directly. Therefore any new auth method can follow the same
pattern and auth can be handled entirely on the backend.

This additionally includes a separate Redis instance for the inference
stack, resolves the pre-commit failure in `main`, and adds missing
dependencies for the inference server.

- Ideally we would use `fastapi-discord` which would handle the Discord
side for us, however there is no version of it which is compatible with
FastAPI versions above `0.84.0`, and we depend on FastAPI `0.88.0`

Some code relating to creating/decoding JWTs is duplicated from existing
backend code and may be possible to move to `oasst_shared`. We also
really need to clean up `main.py` in inference.

After this initial inclusion of users to the server is merged, we will
require two follow-on issues:

- Modify DB schemas and code to associate client actions with DB users
- Lock relevant API calls behind authentication by validating provided
JWTs

The new settings will also need configuring in the environments for the
server to operate properly.

---------

Co-authored-by: Yannic Kilcher <yk@users.noreply.github.com>